Add files via upload
Percolation theory is a powerful framework used to model a wide range of physical and social phenomena, from the flow of fluids through porous materials to the path of electricity in a wire. In this report, we will explore the mathematics behind percolation, seeking to model the paths between two regions of a lattice. We will start by laying the groundwork for percolation analysis, delving into basic examples of percolation on various lattice configurations. We will investigate pathways within square, rectangular, and triangular grids and will explore paths on segments and rotations of triangular lattices. 

Authors: Ashe Daniels, Finlay Dowds, Udayan Nayyar, Dan Phillips